requirements.txt
# HydroFlow Manager - Requirements
# Instalar con: pip install -r requirements.txt

# Base de datos MySQL
mysql-connector-python>=8.0.0

# GUI Framework
customtkinter>=5.0.0
CTkMessagebox>=2.0.0

# Image processing
pillow>=10.0.0

# Plotting and visualization
matplotlib>=3.5.0

# Variables de entorno
python-dotenv>=1.0.0
